moved tests, changed README
