+++ README.md
@@ -1,6 +1,6 @@
 # c1-barefoot-frontend
 
-[![Build Status](https://travis-ci.com/atlp-rwanda/c2-barefoot-frontend.svg?branch=develop)](https://travis-ci.com/atlp-rwanda/c2-barefoot-frontend)
+[![Build Status](https://travis-ci.org/atlp-rwanda/c2-barefoot-frontend.svg?branch=develop)](https://travis-ci.org/atlp-rwanda/c2-barefoot-frontend)
 [![Maintainability](https://api.codeclimate.com/v1/badges/0e0e17e256bddb871c2b/maintainability)](https://codeclimate.com/github/atlp-rwanda/c2-barefoot-frontend/maintainability)
 [![Coverage Status](https://coveralls.io/repos/github/atlp-rwanda/c2-barefoot-frontend/badge.svg?branch=develop)](https://coveralls.io/github/atlp-rwanda/c2-barefoot-frontend?branch=develop)
