+++ __tests__/actions/fetchLocationAction.test.js
@@ -1,7 +1,7 @@
 import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import * as actions from '../../../src/redux/actions/fetchLocationsAction';
-import { locationsPayload } from '../../../dummyData'
+import * as actions from '../../src/redux/actions/fetchLocationsAction';
+import { locationsPayload } from '../../dummyData'
 import moxios from 'moxios'
 
 const middlewares = [thunk]
